Break out augur constants from augur.js for simplified mocking
