+++ packages/augur-ui/src/modules/market/components/core-properties/core-properties.jsx
@@ -12,7 +12,7 @@ import TooltipStyles from "modules/common/less/tooltip";
 import MarketLink from "modules/market/components/market-link/market-link";
 import getValue from "utils/get-value";
 import { dateHasPassed } from "utils/format-date";
-import { constants } from "services/augurjs";
+import { constants } from "services/constants";
 import { Hint } from "modules/common/components/icons";
 import {
   TYPE_REPORT,
