+++ packages/augur-ui/src/modules/market/components/market-basics/market-basics.jsx
@@ -22,7 +22,7 @@ import CommonStyles from "modules/market/components/common/market-common.styles"
 import Styles from "modules/market/components/market-basics/market-basics.styles";
 import SingleSlicePieGraph from "src/modules/market/components/common/single-slice-pie-graph/single-slice-pie-graph";
 import TimeRemainingIndicatorWrapper from "src/modules/market/components/common/time-remaining-indicator/time-remaining-indicator";
-import { constants } from "services/augurjs";
+import { constants } from "services/constants";
 import moment from "moment";
 import { compact } from "lodash";
 import { CategoryTagTrail } from "src/modules/common/components/category-tag-trail/category-tag-trail";
