+++ packages/augur-ui/src/modules/markets/helpers/determine-market-link-type.js
@@ -6,7 +6,7 @@ import {
   TYPE_FINALIZE_MARKET,
   TYPE_CLAIM_PROCEEDS
 } from "modules/markets/constants/link-types";
-import { constants } from "services/augurjs";
+import { constants } from "services/constants";
 import { isEmpty } from "lodash";
 
 export const determineMarketLinkType = (market, loginAccount) => {
