+++ packages/augur-ui/src/modules/reports/actions/load-disputing.js
@@ -1,4 +1,4 @@
-import { augur, constants } from "services/augurjs";
+import { augur} from "services/augurjs";
 import logError from "src/utils/log-error";
 import {
   loadMarketsInfoIfNotLoaded,
@@ -8,6 +8,7 @@ import {
   updateAwaitingDisputeMarkets,
   updateCrowdDisputeMarkets
 } from "modules/reports/actions/update-markets-in-reporting-state";
+import { constants } from "src/services/constants";
 
 export const loadDisputing = (callback = logError) => (dispatch, getState) => {
   const { universe } = getState();
