+++ packages/augur-ui/src/services/augurjs.js
@@ -19,4 +19,3 @@ export const connect = (env, callback = logError) => {
 };
 
 export const augur = new Augur();
-export const { constants } = augur;
